In procedure inifilr, no use keeping values of coefil for all latitudes. Just reuse a one-dimensional array. We can do this by moving computation of coefil in the loops for computation of filtering matrices. We avoid a redundant computation of coefil by putting the computation of direct and inverse matrices inside a same loop on latitudes.
Instead of writing to output file only modfrst, also write associated
latitudes and whether a mode to filter was found.

If a mode was not found (rlamda(modfrst(j)) * cos(rlat(j)) >= 1.) then
the filtering matrix at this latitude is null. So we move this test at
the highest level inside the loop on filtered latitudes. Note that, by
doing so, we do not need to initialize coefil at 0 any longer.