Removed the option to use PPM3D advection schemes. This looks really bugged. In particular, in ppm3d:      do i=1,IMR*JMR 	CRY(i,2) = DTDY*V(i,1,k)      end do while cry is declared as: real CRY(IMR,JNP) and: jmr = jnp - 1
Also removed options for Slopes and Prather advection schemes. This
does not look functional. It seems it needs tracers inside tracers.

Also removed allowed values 1 and 2 for iadv in iniadvtrac. These are
not implemented in advtrac.

With the removal of PPM3D, there is no longer any need to distinguish
between hadv and vadv in iniadvtrac.