The program did not work with cycle_diurne set to false. mu0 in physiq, which is supposed to be a cosine, was set to -999.999. So prmu in swu had a value of the order of 1e3. So zrmum1 in sw2s had a value of the order of 1e3. So zrayl in sw2s had a value of the order of 1e15. So ztray and ptauaz in swclr also had a large value. So zcorae at line 138 in swclr had a large negative value, which resulted in overflow at line 138.
This assignment of -999.999 to mu0 dates from somewhere between
revisions 348 and 524 of LMDZ. It was corrected in revision 1068 of
LMDZ with a call to angle which was present in revision 348. However,
procedure angle was removed from LMDZE in revision 22 because it was
not used. Hesitated to bring back angle but, finally, just removed the
option of having no diurnal cycle.