Bug fix in `CMakeLists.txt`: default value of jjm.
Use the same variable name across procedures for column density of
snow at the surface: fsnow.

Remove useless intermediary variable bils in procedure physiq.

In procedure interfsurf_hq, no need to force zfra between 0 and 1: the
result of the computation is already between 0 and 1 since snow is
necessarily greater than or equal to 0.