Return the new surface temperature from clqh
Avoid computing the difference between old and new surface temperature
in clqh and then adding this difference to the old surface temperature
in `pbl_surface`: return the new surface temperature from clqh instead
of the variation of surface temperature. All in all, we spare a local
array of size knon in clqh and two additions of arrays in
`pbl_surface`, in actual argument of stdlevvar and in computation of
ftsol.