Remove cal and beta in `interfsurf_hq`
Change order of dummy arguments in procedure `calcul_fluxs`: move cal
and beta to the end so they can be conveniently called with keywords.

In procedure `interfsurf_hq`, remove intermediary variables cal and
beta. In case `is_ter`, move the call to `interfsur_lim` just before
the update of albedo.