Call fyhyp and `set_lat` instead of phyetat0
In `test_ozonecm`, call fyhyp and `set_lat` instead of
phyetat0. Motivation: the program crashes with gfortran trying to read
the almost-empty file `restartphy_test_ozonecm.nc`. I think it is a
bug of gfortran. The program works with ifort. To circumvent this bug,
we are not reading `startphy.nc`. This is actually clearer because the
only data we needed from `startphy.nc` was rlat.