Simplify and accelerate coefkz2
Do not repeat the same test on nsrf and inversion inside the loop on
vertical levels. If we are not on ocean then we do not even need to
call coefkz2 and compute a max in `coef_diff_turb`. (Note that coefm
and coefh computed by coefkz are necessarily >= 0.) In coefkz2 we can
invert the loops on i and k so the test on inversion is done only once
for all vertical levels.