Recompute zc and zd from previous time step
Instead of using the saved arrays. Following LMDZ revision 996. As I
understand it, using the saved arrays was a problem for two
reasons. First, it assumed that the surface fractions did not change,
so it required potential surface fractions in `pbl_surface`. Second,
zc and zd were not written to restart file.