Do not write dtvr to start file
Do not write dtvr to an element of `tab_cntrl` in `restart.nc` and do
not read it from `start.nc`. This was only used to print a warning if
different from value chosen in namelist. It does not make much sense
to have a time step in the description of the initial state. Removing
dtvr from start file means we no longer need to call `conf_gcm` in
ce0l and `test_inter_barxy`.