Redefine phis as geopotential in `grid_noro`
Redefine phis as surface geopotential, instead of surface geopotential
height, in `grid_noro`. I did not like that the same name contained
different quantities in `grid_noro` and outside `grid_noro`. Note that
this change does not cost anything in quantity of computation. On the
contrary, we spare the multiplication by g for longitude iim + 1.