Promote phis to variable of module `grid_noro_m`
phis is defined only in `grid_noro` and in dynetat0 where it is just
read from `start.nc`, it is never modified anywhere else. It was
trinkling up and down the stack of calls in ce0l and gcm without being
used in most of the stack. So promoting it as a variable of module
`grid_noro_m` clarifies and simplifies the programs. Note that this
commission does not compile: it remains to do something in dynetat0
where we can no longer read phis.