Move the modification of zgam to `grid_noro`
Move the modification of zgam from orosetup to `grid_noro`. This was
very weird in orosetup. zgam was the only modified variable among
variables coming from `grid_noro`. Furthermore, after the first
execution of orosetup, the modification in orosetup was done again
uselessly at each execution of orosetup: the max operation
did not change anything. And it was repeated uselessly at each
continuation run of gcm.