Change variable `pa` to a named constant
Move variable `pa` from module `dynetat0_chosen_m` to named constant
in module disvert. So we are reverting in part commit bf8abf85. `pa`
was not input at run-time and was never changed. It is never changed
in LMDZ for Earth either (though it is input at run-time in LMDZ for
other planets).