Create intermediary variable zufi
Create intermediary variable zufi for interpolated unpacked zonal
wind. The motivation is to have the same treatment for zonal and
meridional wind. An intermediary variable for interpolated unpacked
meridional wind was created in commit 698e725b. Instead of doing the
present commit, I hesitated to revert commit 698e725b. The advantage
of defining intermediary variables for interpolated unpacked wind is
clarity. The drawback is an additional consumption of main memory for
two three-dimensional arrays, for the duration of calfis.