Do not compute age and albedo of snow on land-ice
Albedo of snow on land-ice is not used. We do not set agesno to 0 on
the landice fraction, we just leave the value which was already
there. We can just leave the value which was already there because we
know the landice fraction does not change during a run. This commit
changes the value of variable AGESNO in `restartphy.nc` on the landice
fraction: 0 instead of something else.