Move the computation of albsol from `pbl_surface` to physiq
Move the computation of albsol from `pbl_surface` to physiq. So we are
reverting part of commit 1eed7b6b. Motivation is to avoid the double
computation of the same value of albsol in physiq, at a time step
where radlwsw is called and at the following time step in
`pbl_surface`. Also, it is clearer that albsol is modified in only one
procedure. Note that this commit does not change the restart files but
it changes the value of NetCDF variable albs in history file because
the value written is now the value at the end of the time step,
corresponding to the value of falbe updated by `pbl_surface`. It is
probably better that the values of albsol and falbe in the history
file are consistent.